Plan 9 from Bell Labs 1994-11-24

+++ port/devmnt.c
@@ -27,7 +27,7 @@ struct Mntalloc
 	Mnt	*list;		/* Mount devices in use */
 	Mnt	*mntfree;	/* Free list */
 	Mntrpc	*rpcfree;
-	int	id;
+	ulong	id;
 	int	rpctag;
 }mntalloc;
